FEAT: launch embedding models
